Bugfix exposing LogFile could cause client logging to stop after rotation
This is because the LogFile that the client could have pointed to is
closed after rotation.

LogFile was exposed to allow client to redirect their stdout and stderr
to the file but this buggy when the file is rotated - client's stdout
and stderr  would still point to the closed file handle.